Update line chart storybook story
