+++ package.json
@@ -389,6 +389,7 @@
     "@storybook/addon-controls": "^6.5.9",
     "@stripe/react-stripe-js": "^1.8.0-0",
     "@stripe/stripe-js": "^1.29.0",
+    "@types/react-resizable": "^3.0.2",
     "@visx/annotation": "^2.10.0",
     "@visx/axis": "^2.11.1",
     "@visx/glyph": "^2.10.0",
@@ -457,6 +458,7 @@
     "react-elm-components": "^1.1.0",
     "react-focus-lock": "^2.7.1",
     "react-grid-layout": "1.3.0",
+    "react-resizable": "^3.0.4",
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-router-dom-v5-compat": "^6.3.0",
